% Generated by roxygen2 (4.0.2): do not edit by hand
\name{src_JDBC}
\alias{src_JDBC}
\alias{tbl.src_JDBC}
\title{Connect to Netezza database with a Netezza JDBC driver}
\usage{
src_JDBC(driver, dbname = NULL, host = NULL, port = NULL,
  password = NULL, url = NULL, ...)

\method{tbl}{src_JDBC}(src, from, ...)
}
\description{
Use \code{src_JDBC} to connect to Netezza database with JDBC driver.
}

